Extracted formValues into separate files + simplified page box padding
