+++ src/pages/addTrade/autoUpdateUnitPrice.tsx
